requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sentence-transformers
faiss-cpu
pypdf
google-generativeai
python-dotenv
slowapi